Format all Markdown files with `dprint`
This is the result of running `dprint fmt` after making two temporary
changes:

- removing `src/` from the list of excluded directories,
- removing the `rustfmt` configuration.

The effect of the first change is that we normalize the formatting of
all Markdown files. The effect of the second is that we avoid
reformatting code blocks in the Markdown files. I want to reformat the
Rust code in a follow-up step since some of the changes seem
unnecessary or undesirable.

Thanks to mdbook-i18n-helpers 0.2, the POT file is nearly unchanged
after this, meaning that all existing translations remain valid. A few
messages were changed because of stray whitespace characters:

     msgid ""
     "Slices always borrow from another object. In this example, `a` has to remain "
    -"'alive' (in scope) for at least as long as our slice. "
    +"'alive' (in scope) for at least as long as our slice."
     msgstr ""

I have not updated CI to require perfectly formatted files. I think
doing so might be annoying since it makes it much harder to submit
small fixes directly from the GitHub editor. So for now, I suggest
that we simply run `dprint` by hand every few weeks.

If dprint/check#11 is fixed so that `dprint` can annotate the lines
that need fixing directly, then I think we can consider more strict
formatting checks.